Factor out FieldUInt
