+++ algebra/primefield/src/lib.rs
@@ -54,9 +54,10 @@ mod proth_field;
 #[cfg(feature = "use_rand")]
 mod rand;
 mod traits;
+mod uint;
 
 // Generic field implementation
-pub use field::{FieldParameters, FieldUInt, PrimeField};
+pub use field::{FieldParameters, PrimeField};
 pub use zkp_u256::MontgomeryParameters;
 
 // The smallest 252-bit Proth field
@@ -72,6 +73,8 @@ pub use zkp_u256 as u256;
 // TODO: Create a prelude module that contains all the useful ones
 pub use traits::{FieldLike, RefFieldLike, Root, SquareRoot};
 pub use zkp_u256::{AddInline, Inv, MulInline, NegInline, One, Pow, SquareInline, SubInline, Zero};
+
+pub use uint::FieldUInt;
 
 // Std/no-std imports
 #[cfg(not(feature = "std"))]
